refactor(run): extract recall_index submodule (Wave C2)
Second cut of the run.rs split (3,619 -> 3,436). Moves the cross-session
recall cluster into run/recall_index.rs: recall_disabled, RECALL_INDEX_BROKEN,
recall_index_allowed, mark_recall_index_broken, reprobe_recall,
probe_recall_at_startup, extract_turn_snippets, IndexJob, recall_index_sender,
index_turn_for_recall.

Two-region move: print_rehydrate_outcome sits in the middle of the cluster and
is mission-rehydrate code (not recall) so it stays in run.rs and remains an
external run:: item. Submodule needs only two imports (crate::theme and
crate::ConversationRuntime); everything else was already fully-qualified or a
local use.

Behavior-identical pure code movement (no A/B): re-exports preserve every
run::X path -- pub for reprobe_recall (commands.rs); pub(crate) for
index_turn_for_recall / probe_recall_at_startup / recall_disabled /
recall_index_allowed (commands.rs, tui_worker, and run.rs's own REPL). The
internal-only items (extract_turn_snippets, recall_index_sender, IndexJob,
mark_recall_index_broken, RECALL_INDEX_BROKEN) stay private to the submodule.

whole-crate build green; fmt + clippy (lean + --all-features) -D warnings clean;
cargo test --lib 1035 passed (unchanged). Stacked on #145 (C1).
